Fixed T2I API Call
Originally, this script would error out on the cronjob.

The only difference between the Apify calls and DD API calls,
was that I was originally pulling from word.dream. So the script now
pulls from the variable. This should (hopefully) fix the HTTP Bad Request
error that sometimes happened when a cronjob ran.

Also cleaned up the code a little.